change-structure: Refactor app file structure - Move files to lib directory - Split functionality of CLI class and main loop - Move main loop logic into bin executable - Update README.md
