+++ bin/calculator
@@ -1,5 +1,32 @@
 #!/usr/bin/env ruby
 
-require_relative '../calculator_cli'
+require_relative '../lib/calculator.rb'
+
+def run_cli
+  interrupted = false
+  exit_message = "\nExiting... Goodbye!"
+  cli = CLI.new
+
+  puts "Welcome to the Calculator CLI. The current calculator module is #{cli.current_module}",
+       "To change calculator, input the name of the desired module. Available calculator are: #{AVAILABLE_MODULES.join(', ')}"
+
+  begin
+    loop do
+      print "> "
+      input = gets
+
+      break if input.nil? || input.chomp.downcase == 'q'
+
+      output = cli.process_input(input)
+
+      puts output unless output.nil?
+    end
+  rescue Interrupt
+    interrupted = true
+    puts exit_message
+  end
+
+  puts exit_message unless interrupted
+end
 
 run_cli
