+++ lib/calculator.rb
@@ -1,0 +1,1 @@
+Dir[File.join(__dir__, 'calculator', '*.rb')].each { |file| require file }
