+++ lib/calculator/standard.rb
